Split out HeapStatData and MetricsImpl, moving the allocator template to its own file.
